[multibody] Clean up some helper file structure
Move internal-only struct out of main MbP header.

Add missing cc file for geometry_contact_data.

Fix class/struct confusion in specializations.
