+++ multibody/plant/BUILD.bazel
@@ -91,6 +91,7 @@ drake_cc_library(
         "deformable_model.cc",
         "discrete_update_manager.cc",
         "dummy_physical_model.cc",
+        "geometry_contact_data.cc",
         "make_discrete_update_manager.cc",
         "multibody_plant.cc",
         "physical_model.cc",
@@ -107,6 +108,7 @@ drake_cc_library(
         "discrete_update_manager.h",
         "dummy_physical_model.h",
         "geometry_contact_data.h",
+        "hydroelastic_contact_forces_continuous_cache_data.h",
         "make_discrete_update_manager.h",
         "multibody_plant.h",
         "multibody_plant_discrete_update_manager_attorney.h",
