+++ multibody/plant/geometry_contact_data.cc
@@ -1,0 +1,4 @@
+#include "drake/multibody/plant/geometry_contact_data.h"
+
+DRAKE_DEFINE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_SCALARS(
+    struct ::drake::multibody::internal::GeometryContactData);
