+++ multibody/plant/multibody_plant.h
@@ -43,22 +43,6 @@ namespace drake {
 namespace multibody {
 namespace internal {
 
-// Structure used in the calculation of hydroelastic contact forces.
-template <typename T>
-struct HydroelasticContactForcesContinuousCacheData {
-  explicit HydroelasticContactForcesContinuousCacheData(int num_bodies) {
-    F_BBo_W_array.resize(num_bodies);
-  }
-
-  // Forces from hydroelastic contact applied to the origin of each body
-  // (indexed by MobodIndex) in the MultibodyPlant.
-  std::vector<SpatialForce<T>> F_BBo_W_array;
-
-  // Information used for contact reporting collected through the evaluation
-  // of the hydroelastic model.
-  std::vector<HydroelasticContactInfo<T>> contact_info;
-};
-
 // Data stored in the cache entry for joint locking.
 template <typename T>
 struct JointLockingCacheData {
@@ -119,6 +103,9 @@ class DiscreteUpdateManager;
 // Forward declarations for geometry_contact_data.h.
 template <typename>
 struct GeometryContactData;
+// Forward declarations for hydroelastic_contact_forces_continuous_cache_data.h.
+template <typename T>
+struct HydroelasticContactForcesContinuousCacheData;
 // Forward declarations for plant_model_attorney.h.
 template <typename>
 class MultibodyPlantModelAttorney;
